makesure/FDB-storage-app/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.5
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dataclasses-json==0.6.3
decorator==5.1.1
distro==1.8.0
executing==2.0.1
fastapi==0.104.1
fastapi-health==0.4.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
ipdb==0.13.13
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.341
langchain-core==0.0.6
langsmith==0.0.67
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==0.27.4
packaging==23.2
pandas==2.1.3
parso==0.8.3
prompt-toolkit==3.0.41
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
traitlets==5.14.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
wcwidth==0.2.12
yarl==1.9.3
